Restructure replica_pipeline_test -> pipeline_test
